Base variant of toolbar representation. Moved from fragment -> activity
