+++ app/src/main/java/me/grechka/yamblz/yamblzweatherapp/presentation/activity/MainActivity.java
@@ -1,8 +1,14 @@
 package me.grechka.yamblz.yamblzweatherapp.presentation.activity;
 
 import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
+import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
+import android.view.View;
 import android.widget.TextView;
 
 import com.arellomobile.mvp.MvpAppCompatActivity;
@@ -16,9 +22,16 @@ import me.grechka.yamblz.yamblzweatherapp.presentation.citySearch.CitySearchFrag
 import me.grechka.yamblz.yamblzweatherapp.presentation.settings.SettingsFragment;
 import me.grechka.yamblz.yamblzweatherapp.presentation.weather.WeatherFragment;
 
-public class MainActivity extends MvpAppCompatActivity implements MainView {
+public class MainActivity extends MvpAppCompatActivity
+        implements
+        NavigationView.OnNavigationItemSelectedListener,
+        MainView {
 
     @InjectPresenter MainPresenter presenter;
+
+    private DrawerLayout drawerLayout;
+    private TextView cityAreaHeaderTextView;
+    private TextView cityTitleHeaderTextView;
 
     @ProvidePresenter
     public MainPresenter providePresenter() {
@@ -30,6 +43,35 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         if (savedInstanceState == null) presenter.showWeather();
+
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        drawerLayout = (DrawerLayout) findViewById(R.id.drawer_layout);
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+                this, drawerLayout, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        drawerLayout.addDrawerListener(toggle);
+        toggle.syncState();
+
+        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
+        View navigationHeaderView = navigationView.getHeaderView(0);
+        navigationView.setNavigationItemSelectedListener(this);
+
+        onHeaderInit(navigationHeaderView);
+    }
+
+    private void onHeaderInit(@NonNull View headerView) {
+        View searchView = headerView.findViewById(R.id.fragment_weather_header_cities_search);
+        cityTitleHeaderTextView = (TextView) headerView.findViewById(R.id.fragment_weather_header_city_title);
+        cityAreaHeaderTextView = (TextView) headerView.findViewById(R.id.fragment_weather_header_city_area);
+
+        //searchView.setOnClickListener(v -> showCitySearch());
+    }
+
+    @Override
+    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+        presenter.navigate(item.getItemId());
+        return true;
     }
 
     @Override
@@ -41,7 +83,7 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
     public void showWeather() {
         getSupportFragmentManager().beginTransaction()
                 .replace(R.id.fragment_container, new WeatherFragment())
-                .commitAllowingStateLoss();
+                .commit();
     }
 
     @Override
@@ -49,7 +91,7 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
         getSupportFragmentManager().beginTransaction()
                 .replace(R.id.fragment_container, new SettingsFragment())
                 .addToBackStack(null)
-                .commitAllowingStateLoss();
+                .commit();
     }
 
     @Override
@@ -57,7 +99,7 @@ public class MainActivity extends MvpAppCompatActivity implements MainView {
         getSupportFragmentManager().beginTransaction()
                 .replace(R.id.fragment_container, new AboutFragment())
                 .addToBackStack(null)
-                .commitAllowingStateLoss();
+                .commit();
     }
 
     @Override
